Refactor bastion client and make it proxy everything
