+++ lib/knife-bastion.rb
@@ -1,0 +1,5 @@
+module Knife
+  module Bastion
+    autoload ClientProxy, 'knife-bastion/client_proxy'
+  end
+end
