+++ lib/knife-bastion/version.rb
@@ -1,6 +1,6 @@
 module Knife
   module Bastion
-    VERSION = "1.0.0"
+    VERSION = "1.0.1"
     MAJOR, MINOR, TINY = VERSION.split('.')
   end
 end
